Split binary decoding and encoding into separate files
