+++ dhall/src/syntax/mod.rs
@@ -9,7 +9,7 @@ mod core;
 pub use crate::syntax::core::context;
 pub use crate::syntax::core::visitor;
 pub use crate::syntax::core::*;
-pub use crate::syntax::text::printer::*;
 pub use crate::syntax::text::parser::*;
+pub use crate::syntax::text::printer::*;
 pub mod binary;
 pub mod text;
